backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pynacl==1.5.0
pyjwt[crypto]==2.8.0
redis==5.0.4
supabase==2.4.6
pydantic==2.7.1
pydantic-settings==2.2.1
pytest==8.2.0
httpx==0.27.0
python-dotenv==1.0.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
